refactor: change project's components structure
