+++ src/components/Search/subComponents/SearchPlaceholder.js
