+++ src/components/Slider/Slider.css
